issue #124 - successfully moved all local (direct DB) OFX stuff to ofxapi.local
