--- docs/source/biweeklybudget.ofxupdater.rst
@@ -1,7 +1,0 @@
-biweeklybudget\.ofxupdater module
-=================================
-
-.. automodule:: biweeklybudget.ofxupdater
-    :members:
-    :undoc-members:
-    :show-inheritance:
